+++ GrailDiary/AppDelegate.swift
@@ -12,7 +12,7 @@ extension UTType {
 public extension Logger {
   static let shared: Logger = {
     var logger = Logger(label: "org.brians-brain.grail-diary")
-    logger.logLevel = .debug
+    logger.logLevel = .info
     return logger
   }()
 }
