+++ GrailDiary/ApplicationMimeType.swift
@@ -1,10 +1,4 @@
-//
-//  ApplicationMimeType.swift
-//  GrailDiary
-//
-//  Created by Brian Dewey on 6/1/21.
-//  Copyright © 2021 Brian's Brain. All rights reserved.
-//
+// Copyright (c) 2018-2021  Brian Dewey. Covered by the Apache 2.0 license.
 
 import Foundation
 
@@ -15,4 +9,3 @@ public enum ApplicationMimeType: String {
   /// MIME type for Book
   case book = "application/json;type=Book"
 }
-
